api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
gunicorn==22.0.0
cassandra-driver==3.29.1
redis==5.0.4
pydantic==2.7.1
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.27.0
prometheus-fastapi-instrumentator==7.0.0
celery==5.4.0
